Update Mpr121 project
